refactor(csharp): 重構 ViewOps 邏輯至獨立的 partial class 檔案
