+++ MCP/Core/CommandExecutor.cs
@@ -1753,36 +1753,6 @@ namespace RevitMCP.Core
         }
 
         /// <summary>
-        /// 重新命名視圖
-        /// </summary>
-        private object RenameView(JObject parameters)
-        {
-            Document doc = _uiApp.ActiveUIDocument.Document;
-            IdType viewId = parameters["viewId"]?.Value<IdType>() ?? 0;
-            string newName = parameters["newName"]?.Value<string>();
-
-            if (string.IsNullOrEmpty(newName))
-                throw new Exception("請指定新的視圖名稱");
-
-            View view = doc.GetElement(new ElementId(viewId)) as View;
-            if (view == null)
-                throw new Exception($"找不到視圖 ID: {viewId}");
-
-            using (Transaction trans = new Transaction(doc, "重新命名視圖"))
-            {
-                trans.Start();
-                view.Name = newName;
-                trans.Commit();
-            }
-
-            return new {
-                ViewId = viewId,
-                NewName = newName,
-                Message = $"成功將視圖重新命名為: {newName}"
-            };
-        }
-
-        /// <summary>
         /// 選取元素
         /// </summary>
         private object SelectElement(JObject parameters)
